backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.6
cryptography==41.0.7
geopy==2.4.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
twilio==8.2.2
python-socketio==5.10.0
aiofiles==23.2.1
pillow==10.1.0
deepface==0.0.79
tensorflow==2.15.0
openai==1.3.0
kyber-py>=1.2.0  # pure-Python CRYSTALS-Kyber (ML-KEM-768) post-quantum KEM